+++ src/app/app.tsx
@@ -1,5 +1,5 @@
 import { AppRouter } from '../providers/router-provider';
-import {StoreProvider} from '../providers/store-provider.tsx';
+import {StoreProvider} from '../providers/store-provider';
 
 export default function App() {
   return (
