+++ src/entities/offers/model/index.ts
@@ -1,0 +1,4 @@
+export * from './types.ts';
+export * from './selectors.ts';
+export * from './slice.ts';
+export * from './action.ts';
