+++ src/entities/offers/model/selectors.ts
@@ -1,4 +1,4 @@
-import {State} from "../../../shared/api/api.ts";
+import {State} from '../../../shared/api';
 
 export const getOffers = (state: State) => state.offersList;
-
+export const getOffersCount = (state: State): number => state.offersCount.countOffers;
