+++ src/entities/offers/model/types.ts
@@ -1,0 +1,11 @@
+import {OfferListItemProps} from '../types';
+
+export type OffersListStateProps = {
+  items: OfferListItemProps[] | undefined;
+  isLoading: boolean;
+  isError: boolean;
+}
+
+export type OffersFilterStateProps = {
+  countOffers: number;
+}
