+++ src/entities/offers/ui/OffersList/offers-list.tsx
@@ -2,17 +2,20 @@ import { FC } from 'react';
 import Sorting from '../../../../features/Sorting';
 import { OfferListItemProps } from '../../types';
 import OffersCardItem from '../OffersCardItem';
+import {useOffersFilter} from '../../hooks';
 
-export const OffersList:FC<{data: OfferListItemProps[]; currentLocation: string}> = ({data, currentLocation}) => (
-  <section className="cities__places places">
-    <h2 className="visually-hidden">Places</h2>
-    <b className="places__found">{data.length} places to stay in {currentLocation}</b>
-    <Sorting/>
-    <div className="cities__places-list places__list tabs__content" data-testid="offers-list">
-      {data && data.map((item: OfferListItemProps) => <OffersCardItem key={item.id} {...item}/>)}
-    </div>
-  </section>
-
-);
+export const OffersList:FC<{currentLocation: string}> = ({currentLocation}) => {
+  const filteredOffers = useOffersFilter(currentLocation);
+  return (
+    <section className="cities__places places">
+      <h2 className="visually-hidden">Places</h2>
+      <b className="places__found">{filteredOffers.length} places to stay in {currentLocation}</b>
+      <Sorting/>
+      <div className="cities__places-list places__list tabs__content" data-testid="offers-list">
+        {filteredOffers.map((item: OfferListItemProps) => <OffersCardItem key={item.id} {...item}/>)}
+      </div>
+    </section>
+  );
+};
