+++ src/entities/offers/ui/OffersListEmpty/index.ts
@@ -1,0 +1,1 @@
+export {OffersListEmpty as default} from './offers-list-empty.tsx';
