+++ src/entities/offers/ui/OffersListEmpty/offers-list-empty.tsx
@@ -1,0 +1,10 @@
+export const OffersListEmpty = () => (
+  <section className="cities__no-places">
+    <div className="cities__status-wrapper tabs__content">
+      <b className="cities__status">No places to stay available</b>
+      <p className="cities__status-description">We could not find any property available at the moment in Dusseldorf</p>
+    </div>
+  </section>
+);
+
+
