+++ src/entities/offers/utils/constants.ts
@@ -1,0 +1,1 @@
+export const INIT_COUNT_OFFERS = 0;
