+++ src/features/Tabs/TabsContent/index.ts
