--- src/features/Tabs/model/selectors.ts
@@ -1,2 +1,0 @@
-import {State} from "../../../shared/api/api.ts";
-export const getOffersCount = (state: State): number => state.offersCount.countOffers;
