+++ src/providers/store-provider.tsx
@@ -1,10 +1,8 @@
 import {configureStore} from '@reduxjs/toolkit';
-import React, {FC} from 'react';
 import {Provider, TypedUseSelectorHook, useDispatch, useSelector} from 'react-redux';
-import {createAPI, State} from "../shared/api/api.ts";
-import {offersSlice} from "../entities/offers/model/slice.ts";
-import {offersFilterSlice} from "../features/Tabs/model/slice.ts";
-
+import {createAPI, State} from '../shared/api/api.ts';
+import {FC, ReactNode} from "react";
+import {offersFilterSlice, offersSlice} from "../entities/offers/model";
 
 export const api = createAPI();
 export const store = configureStore({
@@ -23,8 +21,6 @@ export const store = configureStore({
 
 export type RootState = ReturnType<typeof store.getState>;
 export type AppDispatch = typeof store.dispatch;
-
 export const useAppDispatch = () => useDispatch<AppDispatch>();
-
 export const useAppSelector: TypedUseSelectorHook<State> = useSelector;
-export const StoreProvider: FC = ({children}) => <Provider store={store}>{children}</Provider>;
+export const StoreProvider: FC<{children: ReactNode}> = ({children}) => <Provider store={store}>{children}</Provider>;
